casapepe-backend/requirements.txt
flask
flask-cors
requests
pyairtable
pytest
python-dotenv
